lib/io: add function to read user confirmation from standard input
The function ConfirmYesNo display a question to standard output and
read for answer from input Reader for simple yes "y" or no "n" answer.

If input Reader is nil, it will set to standard input.

If "defIsYes" is true and answer is empty (only new line), then it will
return true.