websocket: add server handler for reserved control frame
The handler is called HandlerRsvControl that can be set before running
the server.  Default HandlerRsvControl is nil.

Since the handler expect Frame as parameter, we need to adjust and add
two methods to access Frame's internal fields and methods; which is
exporting NewFrame function and Frame.Pack method, and adding method
Opcode() and Payload() to access Frame's opcode and payload.