http: add function to parse HTTP header only
In WebSocket handshake, we only need to parse the HTTP header only.
Due to TCP buffering, a server can send an HTTP handshake response (with
header only) and the WebSocket wire frame to client in single read.
If we use standard http.ReadResponse, we can't get the rest of packet
(the initial frame).

To fix this issue we create our own HTTP parser, which limited to parsing
header only without any validation.