smtp: make new client to be implicit connect with EHLO
Previously, Connect and Ehlo are separated functions that must be called
by client manually to connect and to send EHLO command to server.

This commit unexport Connect and Ehlo, and call them implicitly when
creating NewClient.

This minimize any command that require Ehlo first, for example when
upgrading connection with STARTTLS and on mail transaction.