dns: add field name servers to ServerOptions
Field NameServers contains list of parent name servers.

Answer that does not exist on local will be forwarded to parent
name servers.  If this is empty, any query that does not have an
answer in local caches, will be returned with response code
RCodeErrName (3).

The name server use the URI format,

	nameserver  = scheme "://" ( ip-address / domain-name ) [:port]
	scheme      = ( "udp" / "tcp" / "https" )
	ip-address  = ( ip4 / ip6 )
	domain-name = ; fully qualified domain name

The domain-name MUST only used if scheme is "https".

Example,

	udp://1.1.1.1
	tcp://192.168.1.1:5353
	https://cloudflare-dns.com/dns-query

This is the first part of future commits that will forward DNS query
from client to parent name servers.