dns: refactor server to use caches
Previously, we use a handler for server, like server in HTTP library, to
serve query from client based on request.  The idea was to make server
more generic which allow custom cache or request handling.
This is make sense if DNS response is dynamic, but in real world it is
not.  DNS record actually static, either loaded from file (hosts or zone
file) or from response by recursive server.

This commit remove the server Handler and make all request to be
handled automatically based on whether the query exist on cache or not.