dns: simplify server request using io.Writer
Previously, we need the UDP address of client to write back response to
client on UDP connection and http ResponseWriter to write back response
to client of DoH connection.

This changes simplify the sender using io.Writer.

On UDP connection, writer is an instance of UDPClient with
connection reference to UDP server and with peer address.
On TCP connection, writer is a TCP connection from accept.
On Doh connection, writer is http ResponseWriter.