http: allow serving directory with slash
Previously, if a directory contains "index.html" and client request the
directory with slash, server will return 404.  For example,
"/a/index.html" accessed through "/a/" will return an error but "/a" or
"/a/index.html" will return the content of "index.html".

This commit fix this issue by checking if memfs.Get return an
os.ErrNotExist or not.  If its ErrNotExist, concat the request path
with "index.html".