memfs: refactoring go generate file to use type from memfs
Previous Go generated file use their own Node and provide single function
Get() to get the node based on path.  This method does not work when
using it directly with memfs. For example, on server that contains
instance of memfs and want to use the output from GoGenerate, it would
need to add their own wrapper for Get() and memfs.Get().

It should be more simple than that.

This change add global variable GeneratedPathNode to memfs package, that
will be set by file from go generate.  If the GeneratedPathNode is not
nil, memfs will use it as internal mapping of path and node.