memfs: add parameter to make reading file content become optional
Previously, when a directory is mount-ed, all the content of file will
be mapped to memory as long as its size less than MaxFileSize.

The new parameter "withContent" allow the file content to be mapped or not
mapped to memory.  If its false, the MemFS will behave as directory tree.