memfs: add method to update node content and information
The method will update node content and/or information in memory
based on new file information.

This method only check if the node name is equal with new file name, but
it's not checking whether the node is part of memfs (node is parent or
have the same root node).