ini: add method to add, set, and unset variable
The Add() method will add new key and value to the last item in section
and/or subsection.

The Set() method set the last variable's value in section-subsection that
match with the key.

The Unset() method will remove the last variable's in section and/or
subsection that match with the key.