ini: add method Vars that return all variables as map
Given a section name and/or subsection name, Vars() method will return
all variables as map of key and value.

If there is a duplicate in key's name, only the last key value that will
be store on map value.

This method is a shortcut that can be used in templating.