ini: create new section or variable if not exist on Set
Previous behaviour of Set() method will return false if the section
or subsection of key to be set not found on database.

This commit change the behaviour of Set().  If no section or subsection
found on database, the new section with key-value will be created.
If no key found, the new key-value will be added to the specific section.