dns: use channel to stop all forwarders
Previously, we use a boolean condition to stop all forwarders by setting
it to false.  But, this method does not work because using select
statement with single case, will block the process.

This change, allocated a new boolean channel for each forwarders, that
is stored on server fields and when we need to stop or restart all
forwarders, we send the boolean value "true" to each channel.