lib/http: make the request body always available even after ParseForm()
Previously, if the request type is query, form, or JSON, we call the
ParseForm() to let the http.Request read the Body POST form data and fill
the Form and/or PostForm fields.  This  method will cause the request
Body will become empty since its already read and closed.

One of use case of POST with form data is to check the integrity of POST
body using checksum, which is not possible using only ParseForm().

This commit read all the body first into reqBody and recreate the request
Body back using ioutil.NopCloser and bytes.Buffer.