http: add type node and route to handle Endpoint with key
Most of HTTP server allow binding key in the routing, which in turn
allow pretty REST API in URL.  For example, to delete a resource with
unique ID, we can register an Endpoint using HTTP method DELETE with Path
set to  "/resource/:id".  The ":id" is the key.  Its value will be
available in the http.Request's Form field using request.Form.Get("id").