dns: refactor the forwarders logic to handle re-connect gracefully
Previously, if the host connection got disconnected, the forwarders will
closed and recreate.  If client creation is failed it would call
log.Fatal which cause the program to panic.

Instead of using Fatal, we loop and use time.Sleep waiting for host
connection to be active again, every 5 seconds, while still watching
whether the server has been restarted with new nameservers.

Each forwarder now has tag in the following format "<tipe>-<id>-<asWhat>".
The tipe indicate the tipe of forwarders, its either UDP, TCP, DoH, or
DoT.  The id is the number of forwarder.  The asWhat specify the forwarder
either as primary or as fallback.