dns: fix handling server without fallback forwarder
If the ServerOptions' fallback nameservers (FallbackNS) is empty,
the parameter fallbackq on method runXxxForwarder will be nil and this
cause the function incForwarder() will not executed and the processQueue()
will not sent the request to primary queue.

This change introduce boolean parameter to runXxxForwarder which if its
true the incForwarder() will be executed, and allow request to be send
to primary queue without any fallback name servers defined.