sql: add method to migrate database schema
The Migrate method migrate the database using list of SQL files inside
a directory.
Each SQL file in directory will be executed in alphabetical order based on
the last state.

The state of migration will be saved in table "_migration" with the SQL
file name that has been executed and the timestamp.