dns: disable serving DoH and DoT if port is not set (0)
Previously, if HTTPPort or TLSPort is not set (0) we set it to default
value and keep serving the DNS over HTTP and TLS.  Sometimes, user may
want to serve DNS over UDP only.

This changes disable serving DNS over HTTP (DoH) and TLS (DoT) if port
is not set, or 0.