Revert "http: print the endpoint call error only if debug value is 3 or greater"
This reverts commit 4ce1553629b9e944a0bd99b61f2c1ba545bacde0.

Reason for revert: the log is required to debug a system error, the one
that does not come from user predefined log.

For example, if we failed to encode a data with gob or json, the error
that will returned is only "500 Internal server error", we did not know
the actual error.