dns: add method to create Message from hostname and list of addresses
The NewMessageAddress create new DNS message for hostname that contains
one or more A or AAAA addresses.
The addresses must be all IPv4 or IPv6, the first address define the
query type.
If hname is not valid hostname or one of the address is not valid IP
address it will return nil.