paseto: change the Unpack return type to PublicToken
Previously, UnpackHTTPRequest and Unpack methods return the data inside
the token and footer.

This changes, return the unpacked token as a whole, include the JSON
token, the data inside JSON token, and the  JSON footer.

In this way, user can do additional process to the request token that
they receive, for example, to log the request.