memfs: refactoring, allow multiple instances of memfs in single package
Previously, the generated Go code from memfs can be used only once
on the package that use it.  For example, if we have two instances of
memfs.MemFS and both of them call GoGenerate(), when we load them back
again only the last one will be active and set the global variable
memfs.GeneratedPathNode.

This changes refactoring on how we use memfs by storing the
generated path node into variable that is defined by user and pass
them to New Options.

This changes affect package http and io.