dns: change the SOA field in ZoneFile to non-pointer
The reason we use pointer for SOA record is so we can save them only
if its not nil.  A nil SOA record cause the generated JSON become "null"
and this is make the consumer of ZoneFile do more work, checking for the
null and initialize it when required.

This changes, make the SOA record to non-nil and we save the record only
if the MName field is not empty.