lib/clise: new package that implement circular slice
A circular slice is a slice that have fixed size.
An append to slice that has reached its length will overwrite and start
again from index 0.

For example, a clise with size 5,

	c := clise.New(5)
	c.Push(1, 2, 3, 4, 5)
	fmt.Printf("%v\n", c.Slice()) // [1 2 3 4 5]

If we push another item, it will overwrite the first index,

	c.Push(6)
	fmt.Printf("%v\n", c.Slice()) // [6 2 3 4 5]

See the examples for usage of the package.