all: use the ciigo binary instead of library to generate docs
Previously, we use the ciigo as library to convert and server the
internal documentation.  This cause circular dependencies in go.mod
where ciigo depends on share module and share module depends on ciigo.

There is nothing wrong with that and everything seems working fine, until
we need to use "replace" to test the changes on share module from ciigo
module.

This changes remove the use of ciigo library through
internal/cmd/docs-serve and internal/cmd/docs and replace them by
calling the ciigo binary directly.