ini: refactoring unmarshal function
Previously, we use the passed struct as reference for unmarshalling
the ini text into struct. This model on unmarshaling does not works well
if we want to unmarshal to slice of struct.

This commit changes the way the unmarshal works by iterating over
section and variables from parsed ini text file and do a lookup on the
passed struct to find field that match the section, subsection and key.
With this model we can unmarshal a section-subsction into struct
or slice of struct.