http: add fields for paging on GenericResponse
If the response is paging, contains more than one item in data, one
can set the current status of paging in field Offset, Limit, and Count.

The Offset field contains the start of paging.
The Limit field contains the maximum number of records per page.
The Count field contains the total number of records.