http: refactoring parameters on Callback and CallbackErrorHandler
Previously, the parameters to Callback has three types:
the http.ResponseWriter, *http.Request, and []byte for response body.
Not only the type names are long, there is no information on the
registered Endpoint on the receiver of Callback.

This changes wrap the three parameters into single type EndpointRequest
with addition field Endpoint, which contains the registered Endpoint.

On the CallbackErrorHandler we also have three parameters, but instead
of request body we have an error.

This changes store the error for CallbackErrorHandler inside
EndpointRequest.Error field.