ssh: use the SSH agent to authenticate on NewClient
Previously, the client use the IdentityFile on ConfigSection to
authenticate with the remote SSH server.  In case the IdentityFile is
encrypted with passphrase, each time NewClient is invoked it will ask
for passphrase.

To minimize inputing the passphrase, we check if current use run the
SSH agent through SSH_AUTH_SOCK environment variable and use the agent
to generate signers.

This method need manual key added by user to SSH agent before calling
NewClient to make it works.