lib/ssh: move the config parser to subpackage "config"
There are two reasons for moving to sub-package.  First, the code for
parsing the ssh_config(5) take almost 99% of the lines in the ssh package.
Second, in case we want to submit the code to upstream,
golang.org/x/crypto, we need the package to be independent, less external
dependencies as possible.