lib/sql: do not run migration if the last file not exist on the list
Previously, if the last migrated file name not found on the migration
directory, we start executing migration start from the first file.

This changes this behaviour to not run any migration at all.
Since we cannot return it as an error, we only log it.  In the future
we may return it.