lib/memfs: fix symbolic link with different name
Previously, if file is symbolic link and has different name with their
original file, it will return an error when we tried to open the file

  parentpath/filename: no such file or directory

because we use the original file name, not the symlinked file name.

This commit fix this issue by not replacing the original FileInfo for
symlink but by setting only the size and mode.