lib/memfs: fix empty file not being added to tree
Previously, we did not check if the file size is 0 before reading the
content or updating the content type, which cause the read on file
return io.EOF and the file not added to caches.

This commit fix this issue by checking for zero file size and for
io.EOF when reading the file content.