lib/memfs: change the MarshalJSON to always return an object
Previously, MarshalJSON on memfs will return an object of map
of all PathNodes and on Node it will return an object.

This changes make it the JSON response consistent. If its directory
it will return the node object with its childs, without "content".
If its file, it will return the node object with content.

While at it, use single "mod_time" with value is epoch and return
the node ContentType as "content_type".