lib/io: fix NewWatcher when called DirWatcher
When NewWatcher called from DirWatcher's Start(), it will called NewNode
with nil parent parameter.  If the parent parameter is nil on NewNode
the SysPath of new node will be set to the FileInfo.Name() instead of
full or relative path based on current working directory.

Any operation using new node SysPath will failed because the path
does not exist or reachable from current directory.

For example, let say we have the following directory tree,

  testdata
   |
   +--- A
   |
   +--- B

We then set DirWatcher Root to "testdata" from current directory.
The DirWatcher Start then iterate over all child of "testdata" directory,
and call NewWatcher("testdata/A", ...). On the NewWatcher, it will
call NewNode(nil, FileInfo, -1).  Now since the parent is nil,
the Node.SysPath will be set to FileInfo.Name() or base name of the file,
which is "A".

Later, when node content need to be read, ioutil.ReadFile("A") will
fail because the path to "A" does not exist on current directory.

This fix require to force the parameter "parent" on NewNode to be
required.