lib/memfs: export the Node Update method
The Update method update the node metadata or content based on new
file information.
It accept two parameters: the new file information, newInfo, and
maximum file size, maxFileSize.

The newInfo parameter is optional, if its nil, it will read the file
information based on node's SysPath.

The maxFileSize parameter is also optional.
If its negative, the node content will not be updated.
If its zero, it will default to 5 MB.

There are two possible changes that will happen: its either change on
mode or change on content (size and modtime).
Change on mode will not affect the content of node.