lib/io: separate FileState for updated mode and content
Previously, when content of file being watched is modified, it will
trigger the callback with State set to FileStateModified.  When the
mode of file is modified, it will also trigger the callback with
the same state.

This changes separated those state into two kind: FileStateUpdateMode
for update on file mode, and FileStateUpdateContent for update on
file content.