lib/websocket: fix test for client fragmentation
In the test cases we send two unmasked frames.  Since the client
to server must be masked, but in this test we did not mask it,
when the server received first frame the connection may get closed
before we can send the second unmasked frame.

To detect this, we check if the error from send() is "broken pipe"
or not.  If its true then test case stopped.