lib/memfs: fix parent empty directory not removed
Use case:

	x
	x/y

If x/y is empty, and x processed first, the x will
not be removed.

This commit fix this, by sorting the paths in descending order first
to make empty parent removed clearly.  In above case the order
or check become,

	x/y
	x

While at it, update an example of New to give preview of input and
what the expected output for certain operations.