lib/memfs: simplify createRoot and scanDir
The createRoot should not open the file, only call to os.Stat.

The scanDir should only accept the node to be scanned and the call to
Open and Close should be inside the method of scanDir not outside of it.

While at it, prefix all returned errors with method names.