lib/bytes: refactoring and cleaning up the bytes package
The bytes package, and many other packages in this module, is written
when I still learning and using the Go language for master thesis.
Some of the code, like function signature, does not follow the
Go idiom, at least not how the Go source code looks like.

A breaking changes,

* WriteUint16 and WriteUint32 accept slice only without pointer.
  There is no need to pass slice as pointer to function if we want
  to modify the content, as long as the backing storage is not
  changed.

Bug fixes,

* PrintHex: fix print layout on the last line
* ReadHexByte: fix possible index out of range
* SkipAfterToken return -1 and false if no token found, as promised
  in the comment, instead of the length of text.

We move all unit test to example so we have test and example in the
documentation at the same time.

This changes make all test coverage 100%.