lib/totp: make the New to accept only hash with SHA1, SHA256, or SHA512
Previously, the first parameter to New is a function that return
hash.Hash.  This signature can be misleading, because md5.New also
can return hash.Hash but not usable in TOTP.

This changes update the New function signature to accept defined
type that can be set to SHA1, SHA256, or SHA512.