lib/ssh: use agent defined by config or from environment variable
Previously, we only check the environment variable SSH_AUTH_SOCK
to decide whether the client identity will fetched from agent or
from private key file.

This changes use the GetIdentityAgent from config Section to derive the
path to unix socket.  It will return an empty string if the IdentityAgent
is set to "none" in the Section or when SSH_AUTH_SOCK is not exist or
empty.