lib/dns: remove the fallback name servers (NS) from server options
The original idea of fallback NS is to send the query to the one define
to original resolv.conf, instead of using the one defined by user in
ServerOptions NameServers, when an error occured.

But, most of error usually caused by network (disconnected, time out),
so re-sending query to fallback NS does not have any effect if the network
it self is not working.

This changes remove the unnecessary and complex fallback NS from
server.