lib/dns: make the TCP forwarders as complementary of UDP
The TCP forwarders only active when client send the DNS request as TCP.
When the server receive that request it should also forward the request
as TCP not as UDP to prevent the truncated response.

Another use case for TCP is when the response is truncated, the client
will send the query back through TCP connection.  The server should
forward this request using TCP instead of UDP.