lib/dns: add init method to RDataSOA
The ini method initialize the SOA record by setting fields to its default
value if its empty.

The default value for RName is "root", for Serial is current epoch,
for Refresh is 1 day, for Retry is 1 hour, and for Minimum is 1 hour.

While at it, update the comments to be more understandable and represent
the type or fields.