lib/dns: refactoring, create type RecordType to represent type of RR
Previously, we use uint16 to represent type for ResourceRecord Type or
Question type.

To make the code more strict, where parameter or return value, must be
expected as record type, we add new type to represent the RR type:
RecordType.

This changes also rename any variable name of QType or qtype to RType
or rtype because QType is misleading. The type defined the
ResourceRecord to be queried not only question.