lib/sql: make the TruncateTable run with cascade and restart identity
On table that contains foreign key, truncate without cascade may cause
the method fail.

Also, since TruncateTable is, and should be only, used on testing, any
identity columns, for example serial, should be reset back to its initial
value. On PostgreSQL this means the truncate table is with
"RESTART IDENTITY".