lib/http: refactoring NewClient to accept single struct
Previously, the NewClient function accept three parameters: serverURL,
http.Header, and insecure.  If we want to add another parameter,
for example timeout it will cause changes on the function signature.

To prevent this changes in the future, we change it now. The NewClient
now accept single struct.

While at it, we add option to set Timeout.

The Timeout affect the http Transport Timeout and TLSHandshakeTimeout.
The field is optional, if not set it will set to 10 seconds.