lib/mlog: make the Outf method always add new line at the end
One of common mistakes when using logging library is to put the new line
"\n" at then of format string, which cause delayed output written
to Stdout (the OS wait for "\n" as signal for printing).

This changes check new line to every call of Outf method and add it if
its not exist.

If the caller need to call Outf multiple times before ending it with
new line, they should handle it manually by storing into temporary
buffer first and call Outf at the end.