lib/smtp: refactoring NewClient to use struct instead of parameters
Previously, to create new client one must pass three parameters to
NewClient function: localName, remoteURL, and insecure.
If we want to add another parameters in the future, it will cause the
function signature changes.

This changes simplify creating NewClient by passing single struct
with new parameters: AuthUser, AuthPass, and AuthMechanism.
If both AuthUser and AuthPass is not empty, the NewClient will
authenticate the connection, minimize number of step on the caller.