lib/memfs: make the GoEmbed re-mount the Root before generate Go file
Previously, we assume that the PathNodes from instance of MemFS is
already up to date.
If PathNodes is initialized by previous GoEmbed and the Root directory
contains new files or directories, that new files does not get
embedded.

This changes fix this issue by force re-mount the whole Root directory
before generating the Go file.