lib/memfs: changes the DirWatcher and Watcher to use channel
Previously, we use a callback model to propagated changes.
This model has its advantages and disadvantages.

The advantages is there is no limit of queue when the changes need to
be propagated to the caller.
The disadvantages of that the watcher needs to wait for callback to
finish before continue processing.
One can run it under goroutine, but it may cause race if the caller does
not handle update properly and it does not guarantee the goroutine
process it in FIFO.
We can see this on the unit test of NewWatcher, we needs to use
sync.WaitGroup to properly check one changes before processing the order.

This commit changes the DirWatcher and Watcher to use channel, like
the one in time.Ticker.