lib/memfs: update the file mode and/or content on DirWatcher
Previously, the DirWatcher only forward the NodeState if the file being
watched is deleted or modified.

This changes the DirWatcher handle it internally.
If the file is deleted it will be removed from internal MemFS instance.
If the file is updated it will update the mode or content of that file
in the MemFS.