lib/http: implement handler to authorized request to Server Memfs
The FSAuthHandler in the ServerOptions define a function that will
be called to each GET request to Server Memfs instance using the value
from the HTTP Request instance.

If the request is not authorized it must return false and the HTTP
response will be set to 401 Unauthorized with empty body.