lib/http: rewrite ServerOptions FSHandler to handle request in general
This changes rename ServerOptions field FSAuthHandler to FSHandler and
rewrite their usage.

The FSHandler define the function to inspect each GET request to Server
MemFS instance.
The node parameter contains the requested file inside the memfs.

If the handler return true, server will continue processing the node
(writing the Node content type, body, and so on).

If the handler return false, server stop processing the node and return
immediately, which means the function should have already handle writing
the header, status code, and/or body.