lib/memfs: change the Watch method to accept struct
Previously, we assume that the list of files being Watch-ed is equal
to list of files being in Includes.
This may not be correct.
For example, we may want to watch "*.ts" files only but did not want
it to be included during GoEmbed.

This changes introduce list of pattern for files to be watched in
the WatchOptions.Watches field.
If this field is empty, only files pass the Includes filter will be
watched.