lib/dns: add method to get the record in HostsFile by name and/or value
The Get method return the first resource record that match with domain
name and/or value.
The value parameter is optional, if its empty, then only the first record
that match with domain name that will be returned.

If no record matched, it will return nil.

While at it, reformat the hosts_file comments with next Go 1.19 format.