all: rewrite all codes to use "var" instead of ":="
Using ":=" simplify the code but we lose the type.  For example,

  v := F()

The only way we know what the type of v is by inspecting the function
F.
Another disadvantages of using ":=" may cause extra variables
allocation where two or more variables with same type is declared
inside body of function where it could be only one.

While at it, we split the struct for test case into separate type.