lib/reflect: add function Tag to simplify lookup on struct's field tag
Given a StructField and the name of tag, return the tag's value and
options inside the tag.
The options is any string after tag's value, separated by comma.
For example, given the following field definition

	F `tag:"name,opt1, opt2"`

It will return (name, [opt1 opt2], true).

If the field is exported but does not have tag, it will return the field
name (as is without converting to lower case) in val with hasTag is
false: (Name, nil, false).

If the field is un-exported it will return empty val with hasTag is
false ("", nil, false).