lib/reflect: add function Unmarshal
The Unmarshal function set the obj value by calling one of the method:
UnmarshalBinary, UnmarshalJSON, or UnmarshalText; in respective
order.

Just like reflect, the obj value must be pointer to initialized variable
(&T) or pointer-to-pointer to uninitialized variable (**T).

If obj implement one of the method, it will return (true, nil) if there is
no error.

If none of the method exist on obj, it will return (false, nil).