lib/http: add function to unmarshal url.Values using tag `form:`
UnmarshalForm read struct fields tagged with `form:` from out as key and
set its using the value from url.Values based on that key.
If the field does not have `form:` tag but it is exported, then it will
use the field name, in case insensitive.

Only the following types are supported: bool, int/intX, uint/uintX,
floatX, string, []byte, or type that implement BinaryUnmarshaler
(UnmarshalBinary), json.Unmarshaler (UnmarshalJSON), or TextUnmarshaler
(UnmarshalText).

A bool type can be set to true using the following string value: "true",
"yes", or "1".

If the input contains multiple values but the field type is not slice,
the field will be set using the first value.

It will return an error if the out variable is not set-able (the type is
not a pointer to a struct).
It will not return an error if one of the input value is not match with
field type.